src/requirements.txt
pycti==6.6.11
ptnad-client==0.2.1
